Added General Movement, General Rotation & Input Control Scripts
With just 2 General systems we can move all the entities on the scene, changing its velocity value or rotation value. This means we can easily swap or add behaviors without having to change the General Systems at all.
Another system manages Input for both players, and using the new Input system, we can swap controls on the fly as well.