Added ScreenWrapping System
Added a screen wrapping system and converted the camera to entity. When the scene starts, we load the screen limits and then inside the system check the position of all entities that have a speedData component against the limits. If they don't pass the check, they are teleported to the other side. 
Had to add HYBRID_ENTITIES_CAMERA_CONVERSION into the Scripting Define Symbols inside the player settings inside project settings so the camera converts to entity automatically.